EQ1-1-FRONTEND
made a simplified fetch function, as well as a student model and a register component for the student but im having trouble wrapping my head around state transfers in react so i havent yet implemented the transfer of data between the front end and the backend.
also, i've setup a proxy on port 8080.